requirements.txt
Django>=3.2,<3.3
djangorestframework>=3.10
requests==2.22.0